create unique keys when autofill is true
