+++ src/components/Marquee/MarqueeSlider.tsx
@@ -1,3 +1,4 @@
+import { withUniqueKeys } from "@/utils/withUniqueKeys";
 import Marquee from "react-fast-marquee";
 import type { MarqueeProps } from "react-fast-marquee";
 
@@ -5,16 +6,17 @@ export type MarqueeSliderProps = {
   children: React.ReactNode;
 } & MarqueeProps;
 
-export const MarqueeSlider = ({ children }: MarqueeSliderProps) => {
+export const MarqueeSlider = ({ children, ...props }: MarqueeSliderProps) => {
   return (
     <Marquee
+      {...props}
       autoFill
       speed={50}
       gradient
       gradientColor="#0e1011"
       className="h-28"
     >
-      {children}
+      {withUniqueKeys(children)}
     </Marquee>
   );
 };
